Update on domain classes
