+++ rest_api/src/main/java/tqs/cloudit/domain/rest/User.java
@@ -1,0 +1,58 @@
+package tqs.cloudit.domain.rest;
+
+/**
+ *
+ * @author aspedrosa
+ */
+public class User {
+
+    private String username;
+
+    private String password;
+
+    private String name;
+
+    private String email;
+
+    private String type;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+}
