+++ rest_api/src/main/java/tqs/cloudit/repositories/UserRepository.java
@@ -1,0 +1,13 @@
+package tqs.cloudit.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+
+import tqs.cloudit.domain.persistance.User;
+
+/**
+ *
+ * @author aspedrosa
+ */
+public interface UserRepository extends CrudRepository<User, Long> {
+    
+}
